Extract knownRequires to its own file, preparing to be autogenerated
